refactor(cluster): split distributor local deployment tests
