+++ src/agilab/core/test/test_agi_distributor.py
@@ -3689,183 +3689,6 @@ def test_background_job_manager_uses_subprocess_and_real_directories_only(monkey
 
 
 @pytest.mark.asyncio
-async def test_deploy_local_worker_non_source_flow(monkeypatch, tmp_path):
-    app_path = tmp_path / "app"
-    app_path.mkdir(parents=True, exist_ok=True)
-    (app_path / "pyproject.toml").write_text("[project]\nname='demo-app'\n", encoding="utf-8")
-
-    agi_env_root = tmp_path / "agi_env"
-    (agi_env_root / "src" / "agi_env" / "resources").mkdir(parents=True, exist_ok=True)
-    (agi_env_root / "src" / "agi_env" / "resources" / "sample.txt").write_text("x", encoding="utf-8")
-    agi_node_root = tmp_path / "agi_node"
-    agi_node_root.mkdir(parents=True, exist_ok=True)
-
-    cluster_pck = tmp_path / "cluster_pck"
-    (cluster_pck / "agi_distributor").mkdir(parents=True, exist_ok=True)
-    (cluster_pck / "agi_distributor" / "cli.py").write_text("print('cli')", encoding="utf-8")
-
-    wenv_abs = tmp_path / "worker_env"
-    wenv_abs.mkdir(parents=True, exist_ok=True)
-    (wenv_abs / "pyproject.toml").write_text("[project]\nname='worker-app'\n", encoding="utf-8")
-
-    AGI.env = SimpleNamespace(
-        is_source_env=False,
-        is_worker_env=False,
-        install_type=1,
-        agi_env=agi_env_root,
-        agi_node=agi_node_root,
-        active_app=app_path,
-        wenv_abs=wenv_abs,
-        wenv_rel=Path("worker_env"),
-        uv="uv",
-        uv_worker="uv",
-        python_version="3.13",
-        pyvers_worker="3.13",
-        envars={},
-        verbose=1,
-        env_pck=agi_env_root / "src" / "agi_env",
-        dataset_archive=tmp_path / "missing.7z",
-        target_worker="demo_worker",
-        post_install_rel="demo.post_install",
-        user=getpass.getuser(),
-        cluster_pck=cluster_pck,
-    )
-    AGI._run_type = "sync"
-    AGI._rapids_enabled = False
-    AGI._module_to_clean = []
-    commands = []
-
-    async def _fake_run(cmd, cwd):
-        commands.append((cmd, str(cwd)))
-        return ""
-
-    async def _fake_build():
-        return None
-
-    async def _fake_uninstall():
-        return None
-
-    monkeypatch.setattr(AGI, "_hardware_supports_rapids", staticmethod(lambda: False))
-    monkeypatch.setattr(AGI, "_build_lib_local", staticmethod(_fake_build))
-    monkeypatch.setattr(AGI, "_uninstall_modules", staticmethod(_fake_uninstall))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "run", staticmethod(_fake_run))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "set_env_var", staticmethod(lambda *_a, **_k: None))
-
-    await AGI._deploy_local_worker(app_path, Path("worker_env"), " --extra pandas-worker")
-
-    assert AGI._install_done_local is True
-    assert any(" add agi-env" in cmd for cmd, _ in commands)
-    assert any(" add agi-node" in cmd for cmd, _ in commands)
-    assert any("threaded" in cmd for cmd, _ in commands)
-
-
-@pytest.mark.asyncio
-async def test_deploy_local_worker_rapids_reuses_cli_and_falls_back_from_localhost_ssh(monkeypatch, tmp_path):
-    app_path = tmp_path / "app"
-    app_path.mkdir(parents=True, exist_ok=True)
-    (app_path / "pyproject.toml").write_text("[project]\nname='demo-app'\n", encoding="utf-8")
-
-    agi_env_root = tmp_path / "agi_env"
-    (agi_env_root / "src" / "agi_env" / "resources").mkdir(parents=True, exist_ok=True)
-    (agi_env_root / "src" / "agi_env" / "resources" / "sample.txt").write_text("x", encoding="utf-8")
-    agi_node_root = tmp_path / "agi_node"
-    agi_node_root.mkdir(parents=True, exist_ok=True)
-
-    cluster_pck = tmp_path / "cluster_pck"
-    (cluster_pck / "agi_distributor").mkdir(parents=True, exist_ok=True)
-    (cluster_pck / "agi_distributor" / "cli.py").write_text("print('cli')", encoding="utf-8")
-
-    wenv_abs = tmp_path / "worker_env"
-    wenv_abs.mkdir(parents=True, exist_ok=True)
-    (wenv_abs / "pyproject.toml").write_text("[project]\nname='worker-app'\n", encoding="utf-8")
-    existing_cli = wenv_abs.parent / "cli.py"
-    existing_cli.write_text("print('existing cli')", encoding="utf-8")
-
-    AGI.env = SimpleNamespace(
-        is_source_env=False,
-        is_worker_env=False,
-        install_type=1,
-        agi_env=agi_env_root,
-        agi_node=agi_node_root,
-        active_app=app_path,
-        wenv_abs=wenv_abs,
-        wenv_rel=Path("worker_env"),
-        uv="uv",
-        uv_worker="uv",
-        python_version="3.13",
-        pyvers_worker="3.13",
-        envars={},
-        verbose=1,
-        env_pck=agi_env_root / "src" / "agi_env",
-        dataset_archive=tmp_path / "missing.7z",
-        target_worker="demo_worker",
-        post_install_rel="demo.post_install",
-        user="another-user",
-        cluster_pck=cluster_pck,
-        logger=SimpleNamespace(warn=lambda *_a, **_k: None),
-    )
-    AGI._run_type = "sync --dev"
-    AGI._rapids_enabled = True
-    AGI._module_to_clean = []
-    commands = []
-    ssh_calls = []
-    env_vars = []
-    copied_cli = []
-
-    async def _fake_run(cmd, cwd):
-        commands.append((cmd, str(cwd)))
-        return ""
-
-    async def _fake_build():
-        return None
-
-    async def _fake_uninstall():
-        return None
-
-    async def _fake_ssh(_ip, cmd):
-        ssh_calls.append(cmd)
-        raise ConnectionError("localhost ssh denied")
-
-    def _fake_copy(src, dst):
-        copied_cli.append((str(src), str(dst)))
-        raise AssertionError("existing cli.py should be reused")
-
-    monkeypatch.setattr(AGI, "_hardware_supports_rapids", staticmethod(lambda: True))
-    monkeypatch.setattr(AGI, "_build_lib_local", staticmethod(_fake_build))
-    monkeypatch.setattr(AGI, "_uninstall_modules", staticmethod(_fake_uninstall))
-    monkeypatch.setattr(AGI, "exec_ssh", staticmethod(_fake_ssh))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "run", staticmethod(_fake_run))
-    monkeypatch.setattr(
-        agi_distributor_module.AgiEnv,
-        "set_env_var",
-        staticmethod(lambda *args: env_vars.append(args)),
-    )
-    monkeypatch.setattr(agi_distributor_module.shutil, "copy", _fake_copy)
-
-    await AGI._deploy_local_worker(app_path, Path("worker_env"), " --extra pandas-worker")
-
-    assert AGI._install_done_local is True
-    assert ("127.0.0.1", "hw_rapids_capable") in env_vars
-    assert ssh_calls and any("demo.post_install" in cmd for cmd in ssh_calls)
-    assert any("demo.post_install" in cmd for cmd, _ in commands)
-    assert any(
-        "uv sync --config-file uv_config.toml --project" in cmd and str(app_path) in cmd
-        for cmd, _ in commands
-    )
-    assert any(
-        "uv sync --python 3.13 --config-file uv_config.toml --project" in cmd
-        and str(wenv_abs) in cmd
-        for cmd, _ in commands
-    )
-    assert not any(
-        "uv sync" in cmd and str(wenv_abs) in cmd and "--extra pandas-worker" in cmd
-        for cmd, _ in commands
-    )
-    assert copied_cli == []
-    assert any(f'python "{existing_cli}" threaded' in cmd for cmd, _ in commands)
-
-
-@pytest.mark.asyncio
 async def test_deploy_remote_worker_non_source_flow(monkeypatch, tmp_path):
     dist_abs = tmp_path / "dist"
     dist_abs.mkdir(parents=True, exist_ok=True)
@@ -4797,439 +4620,6 @@ async def test_prepare_cluster_env_fallback_installer_and_no_download(monkeypatc
     assert any("python install 3.13" in cmd for _, cmd in cmds)
     assert any(item[2] == "wenv" for item in sent)
     assert any("cli.py" in names for _, names, _ in sent)
-
-
-@pytest.mark.asyncio
-async def test_deploy_local_worker_install_type_zero_non_source_covers_dependency_flow(monkeypatch, tmp_path):
-    repo_root = tmp_path / "repo" / "src" / "agilab"
-    env_project = repo_root / "core" / "agi-env"
-    node_project = repo_root / "core" / "agi-node"
-    core_project = repo_root / "core" / "agi-core"
-    cluster_project = repo_root / "core" / "agi-cluster"
-    for project in (env_project, node_project, core_project, cluster_project):
-        project.mkdir(parents=True, exist_ok=True)
-        (project / "pyproject.toml").write_text(
-            "[project]\nname='demo'\ndependencies=['pip>=1']\n",
-            encoding="utf-8",
-        )
-    (env_project / "src" / "agi_env" / "resources").mkdir(parents=True, exist_ok=True)
-    (env_project / "src" / "agi_env" / "resources" / "resource.txt").write_text("x", encoding="utf-8")
-
-    app_path = tmp_path / "app"
-    app_path.mkdir(parents=True, exist_ok=True)
-    (app_path / "pyproject.toml").write_text("[project]\nname='app'\n", encoding="utf-8")
-    (app_path / "src").mkdir(parents=True, exist_ok=True)
-    (app_path / "src" / "Trajectory.7z").write_text("traj", encoding="utf-8")
-    (app_path / ".venv").mkdir(parents=True, exist_ok=True)
-    (app_path / "uv.lock").write_text("lock", encoding="utf-8")
-
-    wenv_abs = tmp_path / "wenv"
-    wenv_abs.mkdir(parents=True, exist_ok=True)
-    (wenv_abs / "pyproject.toml").write_text("[project]\nname='worker'\n", encoding="utf-8")
-    (wenv_abs / ".venv").mkdir(parents=True, exist_ok=True)
-    (wenv_abs / "_uv_sources" / "ilp_worker").mkdir(parents=True, exist_ok=True)
-
-    env_pck = tmp_path / "env_pck" / "agi_env"
-    env_pck.mkdir(parents=True, exist_ok=True)
-    (env_pck.parent / "__editable__.agi_env-demo.pth").write_text("x", encoding="utf-8")
-
-    dataset_archive = tmp_path / "dataset.7z"
-    dataset_archive.write_text("dataset", encoding="utf-8")
-    share_root = tmp_path / "share"
-    sat_path = share_root / "sat_trajectory" / "dataframe" / "Trajectory"
-    sat_path.mkdir(parents=True, exist_ok=True)
-    (sat_path / "a.csv").write_text("x", encoding="utf-8")
-    (sat_path / "b.csv").write_text("y", encoding="utf-8")
-
-    cluster_pck = tmp_path / "cluster_pck"
-    (cluster_pck / "agi_distributor").mkdir(parents=True, exist_ok=True)
-    (cluster_pck / "agi_distributor" / "cli.py").write_text("print('cli')", encoding="utf-8")
-
-    AGI.env = SimpleNamespace(
-        is_source_env=False,
-        is_worker_env=False,
-        install_type=0,
-        agi_env=env_project,
-        agi_node=node_project,
-        agi_cluster=cluster_project,
-        active_app=app_path,
-        wenv_abs=wenv_abs,
-        wenv_rel=Path("wenv"),
-        uv="uv",
-        uv_worker="uv",
-        python_version="3.13",
-        pyvers_worker="3.13",
-        envars={},
-        verbose=1,
-        env_pck=env_pck,
-        dataset_archive=dataset_archive,
-        target_worker="demo_worker",
-        post_install_rel="demo.post_install",
-        user="another-user",
-        cluster_pck=cluster_pck,
-        share_root_path=lambda: share_root,
-        logger=SimpleNamespace(warn=lambda *_a, **_k: None),
-    )
-    AGI._run_type = "sync --dev"
-    AGI._rapids_enabled = True
-    AGI.agi_workers = {"pandas": "pandas-worker"}
-    commands = []
-    ssh_calls = []
-
-    async def _fake_run(cmd, cwd):
-        commands.append((cmd, str(cwd)))
-        return ""
-
-    async def _fake_build():
-        return None
-
-    async def _fake_uninstall():
-        return None
-
-    async def _fake_ssh(_ip, cmd):
-        ssh_calls.append(cmd)
-        raise ConnectionError("localhost ssh denied")
-
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "read_agilab_path", staticmethod(lambda: repo_root))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "run", staticmethod(_fake_run))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "set_env_var", staticmethod(lambda *_a, **_k: None))
-    monkeypatch.setattr(AGI, "_hardware_supports_rapids", staticmethod(lambda: True))
-    monkeypatch.setattr(AGI, "_build_lib_local", staticmethod(_fake_build))
-    monkeypatch.setattr(AGI, "_uninstall_modules", staticmethod(_fake_uninstall))
-    monkeypatch.setattr(AGI, "exec_ssh", staticmethod(_fake_ssh))
-    monkeypatch.setattr(agi_distributor_module, "_agi__version_missing_on_pypi", lambda _p: True)
-
-    await AGI._deploy_local_worker(app_path, Path("wenv"), " --extra pandas-worker")
-
-    manager_toml = (app_path / "pyproject.toml").read_text(encoding="utf-8")
-    assert "pip" in manager_toml
-    assert (wenv_abs / "src" / "demo_worker" / "dataset.7z").exists()
-    assert (wenv_abs / "src" / "demo_worker" / "Trajectory.7z").exists() is False
-    assert (
-        wenv_abs / ".venv" / "lib" / "python3.13" / "site-packages" / "agilab_uv_sources.pth"
-    ).read_text(encoding="utf-8") == "../../../../_uv_sources\n"
-    assert AGI._install_done_local is True
-    assert any(
-        f'add "{env_project}" "{node_project}"' in cmd and str(wenv_abs) in cmd
-        for cmd, _ in commands
-    )
-    assert not any(
-        "add agi-env" in cmd and str(wenv_abs) in cmd
-        for cmd, _ in commands
-    )
-    assert any("config-file uv_config.toml" in cmd for cmd, _ in commands)
-    assert any("run --project" in cmd and "python -m ensurepip --upgrade" in cmd for cmd, _ in commands)
-    assert any("demo.post_install" in cmd for cmd in ssh_calls)
-
-
-@pytest.mark.asyncio
-async def test_deploy_local_worker_preserves_existing_dependency_ranges(monkeypatch, tmp_path):
-    repo_root = tmp_path / "repo" / "src" / "agilab"
-    env_project = repo_root / "core" / "agi-env"
-    node_project = repo_root / "core" / "agi-node"
-    core_project = repo_root / "core" / "agi-core"
-    cluster_project = repo_root / "core" / "agi-cluster"
-    for project in (env_project, node_project, core_project, cluster_project):
-        project.mkdir(parents=True, exist_ok=True)
-    (env_project / "src" / "agi_env" / "resources").mkdir(parents=True, exist_ok=True)
-    (env_project / "src" / "agi_env" / "resources" / "resource.txt").write_text("x", encoding="utf-8")
-    (env_project / "pyproject.toml").write_text("[project]\nname='agi-env'\ndependencies=['pip>=1']\n", encoding="utf-8")
-    (node_project / "pyproject.toml").write_text(
-        "[project]\nname='agi-node'\ndependencies=['scipy==1.16.1']\n",
-        encoding="utf-8",
-    )
-    (core_project / "pyproject.toml").write_text("[project]\nname='agi-core'\ndependencies=[]\n", encoding="utf-8")
-    (cluster_project / "pyproject.toml").write_text(
-        "[project]\nname='agi-cluster'\ndependencies=['pip>=1']\n",
-        encoding="utf-8",
-    )
-
-    app_path = tmp_path / "app"
-    app_path.mkdir(parents=True, exist_ok=True)
-    (app_path / "pyproject.toml").write_text(
-        "[project]\nname='app'\ndependencies=['scipy>=1.15.2,<1.17']\n",
-        encoding="utf-8",
-    )
-    (app_path / "src").mkdir(parents=True, exist_ok=True)
-    (app_path / ".venv").mkdir(parents=True, exist_ok=True)
-    (app_path / "uv.lock").write_text("lock", encoding="utf-8")
-
-    wenv_abs = tmp_path / "wenv"
-    wenv_abs.mkdir(parents=True, exist_ok=True)
-    (wenv_abs / "pyproject.toml").write_text("[project]\nname='worker'\n", encoding="utf-8")
-    (wenv_abs / ".venv").mkdir(parents=True, exist_ok=True)
-    (wenv_abs / "_uv_sources" / "ilp_worker").mkdir(parents=True, exist_ok=True)
-
-    env_pck = tmp_path / "env_pck" / "agi_env"
-    env_pck.mkdir(parents=True, exist_ok=True)
-    (env_pck.parent / "__editable__.agi_env-demo.pth").write_text("x", encoding="utf-8")
-
-    cluster_pck = tmp_path / "cluster_pck"
-    (cluster_pck / "agi_distributor").mkdir(parents=True, exist_ok=True)
-    (cluster_pck / "agi_distributor" / "cli.py").write_text("print('cli')", encoding="utf-8")
-
-    AGI.env = SimpleNamespace(
-        is_source_env=False,
-        is_worker_env=False,
-        install_type=0,
-        agi_env=env_project,
-        agi_node=node_project,
-        agi_cluster=cluster_project,
-        active_app=app_path,
-        wenv_abs=wenv_abs,
-        wenv_rel=Path("wenv"),
-        uv="uv",
-        uv_worker="uv",
-        python_version="3.13",
-        pyvers_worker="3.13",
-        envars={},
-        verbose=1,
-        env_pck=env_pck,
-        dataset_archive=tmp_path / "dataset.7z",
-        target_worker="demo_worker",
-        post_install_rel="demo.post_install",
-        user="another-user",
-        cluster_pck=cluster_pck,
-        share_root_path=lambda: tmp_path / "share",
-        logger=SimpleNamespace(warn=lambda *_a, **_k: None),
-    )
-    AGI._run_type = "sync --dev"
-    AGI._rapids_enabled = False
-    AGI.agi_workers = {"pandas": "pandas-worker"}
-
-    commands = []
-
-    async def _fake_run(cmd, cwd):
-        commands.append((cmd, str(cwd)))
-        return ""
-
-    async def _fake_build():
-        return None
-
-    async def _fake_uninstall():
-        return None
-
-    async def _fake_ssh(_ip, _cmd):
-        raise ConnectionError("localhost ssh denied")
-
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "read_agilab_path", staticmethod(lambda: repo_root))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "run", staticmethod(_fake_run))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "set_env_var", staticmethod(lambda *_a, **_k: None))
-    monkeypatch.setattr(AGI, "_hardware_supports_rapids", staticmethod(lambda: False))
-    monkeypatch.setattr(AGI, "_build_lib_local", staticmethod(_fake_build))
-    monkeypatch.setattr(AGI, "_uninstall_modules", staticmethod(_fake_uninstall))
-    monkeypatch.setattr(AGI, "exec_ssh", staticmethod(_fake_ssh))
-    monkeypatch.setattr(agi_distributor_module, "_agi__version_missing_on_pypi", lambda _p: False)
-
-    await AGI._deploy_local_worker(app_path, Path("wenv"), " --extra pandas-worker")
-
-    manager_toml = (app_path / "pyproject.toml").read_text(encoding="utf-8")
-    assert 'scipy>=1.15.2,<1.17' in manager_toml
-    assert 'scipy==1.16.1' not in manager_toml
-    assert any(
-        f'add "{env_project}" "{node_project}"' in cmd and str(wenv_abs) in cmd
-        for cmd, _ in commands
-    )
-    assert any("sync --project" in cmd for cmd, _ in commands)
-
-
-@pytest.mark.asyncio
-async def test_deploy_local_worker_infers_repo_root_to_avoid_rewriting_source_app(monkeypatch, tmp_path):
-    repo_root = tmp_path / "repo" / "src" / "agilab"
-    env_project = repo_root / "core" / "agi-env"
-    node_project = repo_root / "core" / "agi-node"
-    core_project = repo_root / "core" / "agi-core"
-    cluster_project = repo_root / "core" / "agi-cluster"
-    for project in (env_project, node_project, core_project, cluster_project):
-        project.mkdir(parents=True, exist_ok=True)
-    (env_project / "src" / "agi_env" / "resources").mkdir(parents=True, exist_ok=True)
-    (env_project / "src" / "agi_env" / "resources" / "resource.txt").write_text("x", encoding="utf-8")
-    (env_project / "pyproject.toml").write_text("[project]\nname='agi-env'\ndependencies=['pip>=1']\n", encoding="utf-8")
-    (node_project / "pyproject.toml").write_text(
-        "[project]\nname='agi-node'\ndependencies=['scipy==1.16.1']\n",
-        encoding="utf-8",
-    )
-    (core_project / "pyproject.toml").write_text("[project]\nname='agi-core'\ndependencies=[]\n", encoding="utf-8")
-    (cluster_project / "pyproject.toml").write_text(
-        "[project]\nname='agi-cluster'\ndependencies=['pip>=1']\n",
-        encoding="utf-8",
-    )
-
-    fake_module_file = (
-        repo_root
-        / "core"
-        / "agi-cluster"
-        / "src"
-        / "agi_cluster"
-        / "agi_distributor"
-        / "agi_distributor.py"
-    )
-    fake_module_file.parent.mkdir(parents=True, exist_ok=True)
-    fake_module_file.write_text("# fake module path for repo inference\n", encoding="utf-8")
-
-    app_path = repo_root / "apps" / "builtin" / "demo_project"
-    app_path.mkdir(parents=True, exist_ok=True)
-    (app_path / "pyproject.toml").write_text(
-        "[project]\nname='demo-project'\ndependencies=['py7zr']\n",
-        encoding="utf-8",
-    )
-    (app_path / "src").mkdir(parents=True, exist_ok=True)
-    (app_path / ".venv").mkdir(parents=True, exist_ok=True)
-    (app_path / "uv.lock").write_text("lock", encoding="utf-8")
-
-    wenv_abs = tmp_path / "wenv"
-    wenv_abs.mkdir(parents=True, exist_ok=True)
-    (wenv_abs / "pyproject.toml").write_text("[project]\nname='worker'\n", encoding="utf-8")
-    (wenv_abs / ".venv").mkdir(parents=True, exist_ok=True)
-    (wenv_abs / "_uv_sources" / "ilp_worker").mkdir(parents=True, exist_ok=True)
-
-    env_pck = tmp_path / "env_pck" / "agi_env"
-    env_pck.mkdir(parents=True, exist_ok=True)
-    (env_pck.parent / "__editable__.agi_env-demo.pth").write_text("x", encoding="utf-8")
-
-    cluster_pck = tmp_path / "cluster_pck"
-    (cluster_pck / "agi_distributor").mkdir(parents=True, exist_ok=True)
-    (cluster_pck / "agi_distributor" / "cli.py").write_text("print('cli')", encoding="utf-8")
-
-    AGI.env = SimpleNamespace(
-        is_source_env=False,
-        is_worker_env=False,
-        install_type=0,
-        agi_env=env_project,
-        agi_node=node_project,
-        agi_cluster=cluster_project,
-        active_app=app_path,
-        wenv_abs=wenv_abs,
-        wenv_rel=Path("wenv"),
-        uv="uv",
-        uv_worker="uv",
-        python_version="3.13",
-        pyvers_worker="3.13",
-        envars={},
-        verbose=1,
-        env_pck=env_pck,
-        dataset_archive=tmp_path / "dataset.7z",
-        target_worker="demo_worker",
-        post_install_rel="demo.post_install",
-        user="another-user",
-        cluster_pck=cluster_pck,
-        share_root_path=lambda: tmp_path / "share",
-        logger=SimpleNamespace(warn=lambda *_a, **_k: None),
-    )
-    AGI._run_type = "sync --dev"
-    AGI._rapids_enabled = False
-    AGI.agi_workers = {"pandas": "pandas-worker"}
-
-    commands = []
-
-    async def _fake_run(cmd, cwd):
-        commands.append((cmd, str(cwd)))
-        return ""
-
-    async def _fake_build():
-        return None
-
-    async def _fake_uninstall():
-        return None
-
-    async def _fake_ssh(_ip, _cmd):
-        raise ConnectionError("localhost ssh denied")
-
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "read_agilab_path", staticmethod(lambda: None))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "run", staticmethod(_fake_run))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "set_env_var", staticmethod(lambda *_a, **_k: None))
-    monkeypatch.setattr(agi_distributor_module, "__file__", str(fake_module_file))
-    monkeypatch.setattr(AGI, "_hardware_supports_rapids", staticmethod(lambda: False))
-    monkeypatch.setattr(AGI, "_build_lib_local", staticmethod(_fake_build))
-    monkeypatch.setattr(AGI, "_uninstall_modules", staticmethod(_fake_uninstall))
-    monkeypatch.setattr(AGI, "exec_ssh", staticmethod(_fake_ssh))
-    monkeypatch.setattr(agi_distributor_module, "_agi__version_missing_on_pypi", lambda _p: False)
-
-    await AGI._deploy_local_worker(app_path, Path("wenv"), " --extra pandas-worker")
-
-    manager_toml = (app_path / "pyproject.toml").read_text(encoding="utf-8")
-    assert "py7zr" in manager_toml
-    assert "pip>=1" not in manager_toml
-    assert "scipy==1.16.1" not in manager_toml
-    assert any(
-        f'add "{env_project}" "{node_project}"' in cmd and str(wenv_abs) in cmd
-        for cmd, _ in commands
-    )
-
-
-@pytest.mark.asyncio
-async def test_deploy_local_worker_source_env_branch(monkeypatch, tmp_path):
-    app_path = tmp_path / "app"
-    app_path.mkdir(parents=True, exist_ok=True)
-    (app_path / "pyproject.toml").write_text("[project]\nname='app'\n", encoding="utf-8")
-    wenv_abs = tmp_path / "wenv"
-    wenv_abs.mkdir(parents=True, exist_ok=True)
-    (wenv_abs / "pyproject.toml").write_text("[project]\nname='worker'\n", encoding="utf-8")
-    agi_env = tmp_path / "agi_env"
-    agi_node = tmp_path / "agi_node"
-    agi_cluster = tmp_path / "agi_cluster"
-    for p in (agi_env, agi_node, agi_cluster):
-        p.mkdir(parents=True, exist_ok=True)
-        (p / "dist").mkdir(parents=True, exist_ok=True)
-    (agi_env / "dist" / "agi_env-0.0.1-py3-none-any.whl").write_text("whl", encoding="utf-8")
-    (agi_node / "dist" / "agi_node-0.0.1-py3-none-any.whl").write_text("whl", encoding="utf-8")
-    cluster_pck = tmp_path / "cluster_pck"
-    (cluster_pck / "agi_distributor").mkdir(parents=True, exist_ok=True)
-    (cluster_pck / "agi_distributor" / "cli.py").write_text("print('cli')", encoding="utf-8")
-
-    AGI.env = SimpleNamespace(
-        is_source_env=True,
-        is_worker_env=False,
-        install_type=1,
-        agi_env=agi_env,
-        agi_node=agi_node,
-        agi_cluster=agi_cluster,
-        active_app=app_path,
-        wenv_abs=wenv_abs,
-        wenv_rel=Path("wenv"),
-        uv="uv",
-        uv_worker="uv",
-        python_version="3.13",
-        pyvers_worker="3.13",
-        envars={},
-        verbose=1,
-        env_pck=tmp_path / "env_pck",
-        dataset_archive=tmp_path / "missing.7z",
-        target_worker="demo_worker",
-        post_install_rel="demo.post_install",
-        user=getpass.getuser(),
-        cluster_pck=cluster_pck,
-        share_root_path=lambda: tmp_path / "share",
-        logger=SimpleNamespace(warn=lambda *_a, **_k: None),
-    )
-    AGI._run_type = "sync"
-    AGI._rapids_enabled = False
-    commands = []
-
-    async def _fake_run(cmd, cwd):
-        commands.append((cmd, str(cwd)))
-        return ""
-
-    async def _fake_build():
-        return None
-
-    async def _fake_uninstall():
-        return None
-
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "run", staticmethod(_fake_run))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "set_env_var", staticmethod(lambda *_a, **_k: None))
-    monkeypatch.setattr(AGI, "_hardware_supports_rapids", staticmethod(lambda: False))
-    monkeypatch.setattr(AGI, "_build_lib_local", staticmethod(_fake_build))
-    monkeypatch.setattr(AGI, "_uninstall_modules", staticmethod(_fake_uninstall))
-    monkeypatch.setattr(agi_distributor_module, "_agi__version_missing_on_pypi", lambda _p: False)
-
-    await AGI._deploy_local_worker(app_path, Path("wenv"), " --extra pandas-worker")
-
-    assert AGI._install_done_local is True
-    assert any("pip install -e '" in cmd and str(agi_cluster) in cmd for cmd, _ in commands)
-    assert any("build --wheel" in cmd and str(agi_env) in cmd for cmd, _ in commands)
-    assert any("build --wheel" in cmd and str(agi_node) in cmd for cmd, _ in commands)
-    assert (wenv_abs / "agi_node-0.0.1-py3-none-any.whl").exists()
 
 
 @pytest.mark.asyncio
